+++ src/Host/Worker/Jobs/QueueConsumptionJob.cs
@@ -3,6 +3,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
 using Quartz;
+using ViajaNet.JobApplication.Application;
 using ViajaNet.JobApplication.Infrastructure.Queue;
 
 namespace ViajaNet.JobApplication.Host.Worker
@@ -26,24 +27,24 @@ namespace ViajaNet.JobApplication.Host.Worker
         /// <returns>A <see cref="Task"/> of the current job execution.</returns>
         public async Task Execute(IJobExecutionContext context)
         {
-            if (context.MergedJobDataMap.TryGetValue("queueProvider", out object jobProvider))
+            if (context.MergedJobDataMap.TryGetValue("appService", out object appService))
             {
-                var provider = jobProvider as IQueueProvider;
+                var service = appService as IAnalyticsAppService;
 
-                if (provider == null)
+                if (service == null)
                 {
                     await Console.Out.WriteLineAsync($"Job \"{context.JobDetail.Key}\" started.");
 
                     return;
                 }
 
-                provider.Shift("analytics", async (sender, e) =>
+                /* provider.Shift("analytics", async (sender, e) =>
                 {
                     var json = Encoding.UTF8.GetString(e.Body);
                     var hit = JsonConvert.DeserializeObject<object>(json);
 
                     await Console.Out.WriteAsync("Queue consumption succeeded.");
-                });
+                }); */
             }
             else
             {